Create a SimpleCleaners class and refactor many small cleaner files into the class
