+++ src/cleaners/cleanConcertData.js
@@ -1,9 +1,7 @@
-import { CleanArtists } from './cleanArtists';
-import { Dates } from './Dates';
-const moment = require('moment');
+import { SimpleCleaners } from './SimpleCleaners';
 
-const cleanArtists = new CleanArtists();
-const time = new Dates();
+const moment = require('moment');
+const clean = new SimpleCleaners();
 
 export const cleanConcertData = concerts => {
   concerts = concerts.filter(concert => {
@@ -40,13 +38,12 @@ export const cleanConcertData = concerts => {
       state: show.region_abbr,
       zip: show.postal_code
     };
-    const artists = cleanArtists.clean(show);
+    const artists = clean.artists(show);
     const headlineArtist = artists[0];
     const supportArtists = artists.filter(artist => artist !== artists[0]);
-    const ticketUrl = cleanTickets(links, tickets);
-
+    const ticketUrl = clean.tickets(links, tickets);
     const date = moment(start_time).format('MMM D');
-    const startTime = time.cleanTime(start_time);
+    const startTime = clean.time(start_time);
     const concertData = {
       headlineArtist,
       supportArtists,
@@ -63,20 +60,4 @@ export const cleanConcertData = concerts => {
 
     return [...concertArray, concertData];
   }, []);
-};
-
-const cleanTickets = (links, tickets) => {
-  if (links) {
-    const preferred = links.link.find(ticket => ticket.url.includes('axs.com'));
-
-    if (preferred) {
-      return preferred.url;
-    }
-  }
-
-  if (tickets) {
-    return tickets.link[0].url;
-  }
-
-  return null;
 };
