--- src/cleaners/cleanImage.js
@@ -1,10 +1,0 @@
-import defaultImage from '../images/black-woven.jpg';
-
-export const cleanImage = (artistData, currentImage) => {
-  if (!artistData.artist || !artistData.artist.image[4]) {
-    return currentImage || defaultImage;
-  }
-  const artistImage = artistData.artist.image[4]['#text'];
-
-  return artistImage;
-};
